Adjusting the OpenAi integration
I started with the required restrictions for the open ai request (Only one request per day for each user).

I also created a landing page( it still needs some work ). I ask of the front end team to make the landing page better and to create a popup that shows some static data to distract the user from the waiting time of the ai